Add new feature, refactor code
Add Logger to project, add RandomDataUtils class, refactor code.

+++ src/test/java/com/spribe/config/BaseTest.java
@@ -5,25 +5,38 @@ import com.spribe.api.models.request.DeletePlayerRequest;
 import com.spribe.enums.UserRole;
 import com.spribe.service.PlayerService;
 import com.spribe.utils.ConfigReader;
+import io.qameta.allure.restassured.AllureRestAssured;
 import io.restassured.RestAssured;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeClass;
+import org.testng.annotations.BeforeSuite;
 
 import java.util.List;
 
 public class BaseTest {
 
-    @BeforeClass
+    private static final Logger LOGGER = LogManager.getLogger(BaseTest.class);
+
+    private final int retryCountValue = ConfigReader.getIntProperty("retry_count");
+
+    @BeforeSuite
     public void setup() {
-        RestAssured.baseURI = ConfigReader.getProperty("base_url");
+        LOGGER.info("Retry count value is set to : {}", retryCountValue);
+        RestAssured.filters(new AllureRestAssured());
+        RestAssured.baseURI = ConfigReader.getStringProperty("base_url");
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
     @AfterMethod
     public void clearPlayers() {
         List<String> playerIdsToRemove = PlayerService.getPlayerIdsToRemove();
-        playerIdsToRemove.forEach(player -> PlayerApi.deletePlayerById(UserRole.ADMIN, new DeletePlayerRequest(player)));
-        PlayerService.clearList();
+        if (!playerIdsToRemove.isEmpty()) {
+            playerIdsToRemove.forEach(player -> PlayerApi.deletePlayerById(UserRole.ADMIN, new DeletePlayerRequest(player)));
+            PlayerService.clearList();
+        } else {
+            LOGGER.info("There are no players to remove");
+        }
     }
 
 }
